refactor: move db and error feature to src/utils and moved class definitions to individule files
